rename: WritePostPage -> UploadedVotePage로 폴더 변경
