+++ src/pages/UploadedVotePage/components/Content.tsx
